Add function-specific keyword-arguments to metric functions. This is done, as the same arguments passed via the unspecific kwargs-argument are then further passed for example to both perturb_func and normalise_func.
This commit therefore introduces normalise_func_kwargs, model_predict_kwargs and
perturb_func_kwargs. For the current version the actual behaviour will not be
changed, so passing unspecific kwargs is still working as before, i.e.
they are being passed to normalise_func and perturb_func.
This behaviour is intended to be deprecated and eventually removed in
future versions.

For this feature, the base.Metric class was extended to contain all the
general attributes and preprocessing which was present in each specific
metric. Most of this is now handled by the prepare() method of the
base-class. There is also the base.PerturbationMetric class added, which
handles the perturbation attributes. This way maintainability and
readability is much improved, as fixing issues can potentially be solved
in the base-class only, without needing to go through all of the
metrics.

Each metric now has to implement a process_instance() method, where a
single instance of the input batch is processed and the score is simply
returned. The scores will then be written in the base-class into the last_results
attribute. Addtionally, each metric can implement preprocess() and
postprocess() methods. If the structure of the base-class doesn't fit
this design, it can still be implemented the old way, like seen in the
metric ModelParameterRandomisation, which nevertheless still uses the handy prepare()
method from the base class.